+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.2
+
+add From/TryFrom instances
+add more functions to find values in the LDAPSchema
+change functions to find values in the LDAPSchema to work on anything with an impl TryFrom<KeyStringOrOID>
+add Hash instances
+add functions to return required and allowed attributes for a given objectClass
+
 ## 0.1.1
 
 add serde support